Callbacks

Callbacks are functions that are called at certain points in the training process. They are useful for logging, early stopping, and other tasks. Callbacks are passed to the fit! function as a list, and implement their functionality by implementing a subset of the Hooks.

Available Callbacks

Checkpoints

Tsunami.Checkpointer — Type

Checkpointer(folder = nothing) <: AbstractCallback

An helper class for saving a FluxModule and the fit state. The checkpoint is saved as a BSON file with the name ckpt_epoch=X_step=Y.bson. A symbolic link to the last checkpoint is also created as ckpt_last.bson.

A Checkpointer is automatically created when checkpointer = true is passed to fit!.

If folder is not specified, the checkpoints are saved in a folder named checkpoints in the run directory.

See also: load_checkpoint.

Examples

checkpointer = Checkpointer()
Tsunami.fit!(model, trainer, callbacks = [checkpointer])

Tsunami.load_checkpoint — Function

load_checkpoint(path)

Loads a checkpoint that was saved to path. Returns a namedtuple of the model and the optimizer.

See also: Checkpointer.

Writing Custom Callbacks

See the implementation of Checkpointer and the Hooks section of the documentation for more information on how to write custom callbacks.