FluxModule

Tsunami.FluxModule — Type

FluxModule

An abstract type for Flux models. A FluxModule helps orgainising you code and provides a standard interface for training.

A FluxModule comes with functor already implemented. You can change the trainables by implementing Optimisers.trainables.

Types inheriting from FluxModule have to be mutable. They also have to implement the following methods in order to interact with a Trainer.

Required methods

configure_optimisers(model)

training_step(model, batch, batch_idx)

Optional Methods

validation_step(model, batch, batch_idx)

test_step(model, batch, batch_idx)

training_epoch_end(model, outs)

validation_epoch_end(model, outs)

test_epoch_end(model, outs)

Examples

using Flux, Tsunami, Optimisers

# Define a Multilayer Perceptron implementing the FluxModule interface

mutable struct Model <: FluxModule
    net
end

function Model()
    net = Chain(Dense(4 => 32, relu), Dense(32 => 2))
    return Model(net)
end

(model::Model)(x) = model.net(x)

function Tsunami.training_step(model::Model, batch, batch_idx)
    x, y = batch
    y_hat = model(x)
    loss = Flux.Losses.mse(y_hat, y)
    return loss
end

function Tsunami.configure_optimisers(model::Model)
    return Optimisers.setup(Optimisers.Adam(1e-3), model)
end

# Prepare the dataset and the DataLoader

X, Y = rand(4, 100), rand(2, 100)
train_dataloader = Flux.DataLoader((x, y), batchsize=10)

# Create and Train the model

model = Model()
trainer = Trainer(max_epochs=10)
Tsunami.fit!(model, trainer; train_dataloader)

Tsunami.configure_optimisers — Function

configure_optimisers(model)

Return an optimiser's state initialized for the model. It can also return a tuple of (scheduler, optimiser), where scheduler is any callable object that takes the current epoch as input and returns a scalar that will be set as the learning rate for the next epoch.

Examples

using Optimisers, ParameterScheduler

function Tsunami.configure_optimisers(model::Model)
    return Optimisers.setup(AdamW(1e-3), model)
end

# Now with a scheduler dropping the learning rate by a factor 10 
# at epochs [50, 100, 200] starting from the initial value of 1e-2
function Tsunami.configure_optimisers(model::Model)
    
    function lr_scheduler(epoch)
        if epoch <= 50
            return 1e-2
        elseif epoch <= 100
            return 1e-3
        elseif epoch <= 200
            return 1e-4
        else
            return 1e-5
        end
    end
    
    opt = Optimisers.setup(AdamW(), model)
    return lr_scheduler, opt
end

# Same as above but using the ParameterScheduler package.
function Tsunami.configure_optimisers(model::Model)
    lr_scheduler = ParameterScheduler.Step(1e-2, 1/10, [50, 50, 100])
    opt = Optimisers.setup(AdamW(), model)
    return lr_scheduler, opt
end

Tsunami.test_step — Function

test_step(model, batch, batch_idx)

If not implemented, the default is to use validation_step.

Tsunami.test_epoch_end — Function

test_epoch_end(model, outs)

If not implemented, the default is to use validation_epoch_end.

Tsunami.training_step — Function

training_step(model, batch, batch_idx)

Should return either a scalar loss or a NamedTuple with a scalar 'loss' field.

Tsunami.training_epoch_end — Function

training_epoch_end(model, outs)

If not implemented, do nothing.

Tsunami.validation_step — Function

validation_step(model, batch, batch_idx)

If not implemented, the default is to use training_step. The return type has to be a NamedTuple.

Tsunami.validation_epoch_end — Function

validation_epoch_end(model, outs)

If not implemented, the default is to compute the mean of the scalar outputs of validation_step.