FluxModule

Tsunami.FluxModule — Type

abstract type FluxModule end

Type inheriting from FluxModule have to be mutable. A FluxModule comes with functor already implemented. You can change the trainables by implementing Optimisers.trainables.

FluxModule's subtypes need to implement the following methods:

configure_optimisers(model)

training_step(model, batch, batch_idx)

Optionally also:

validation_step(model, batch, batch_idx)

test_step(model, batch, batch_idx)

training_epoch_end(model, outs)

validation_epoch_end(model, outs)

test_epoch_end(model, outs)

Tsunami.configure_optimisers — Function

configure_optimisers(model)

Tsunami.test_step — Function

test_step(model, batch, batch_idx)

If not implemented, the default is to use validation_step.

Tsunami.test_epoch_end — Function

test_epoch_end(model::MyModule, outs)

If not implemented, the default is to use validation_epoch_end.

Tsunami.training_step — Function

training_step(model, batch, batch_idx)

Should return either a scalar loss or a NamedTuple with a scalar 'loss' field.

Tsunami.training_epoch_end — Function

training_epoch_end(model, outs)

If not implemented, do nothing.

Tsunami.validation_step — Function

validation_step(model, batch, batch_idx)

If not implemented, the default is to use training_step. The return type has to be a NamedTuple.

Tsunami.validation_epoch_end — Function

validation_epoch_end(model::MyModule, outs)

If not implemented, the default is to compute the mean of the scalar outputs of validation_step.