Hooks

Hooks are a way to extend the functionality of Tsunami. They are a way to inject custom code into the FluxModule or into a Callback at various points in the training, testing, and validation loops.

At a high level, and omitting function imputs and outputs, a simplified version of the Tsunami.fit method looks like this:

function fit()
    configure_optimizers()
    
    for epoch in epochs
        train_loop()
    end
end

function train_loop()
    on_train_epoch_start()
    set_learning_rate(lr_scheduler, epoch)

    for batch in train_dataloader
        on_train_batch_start()
        batch = transfer_batch_to_device(batch)
        loss, pb = pullback(m -> train_step(model, batch),  model)
        on_before_backprop()
        grad = pb(1)
        on_before_update()
        update!(opt_state, model, grad)
        on_train_batch_end()
        if should_check_val
            val_loop()
        end
    end
    on_train_epoch_end()
end

function val_loop()
    on_val_epoch_start()
    for batch in val_dataloader
        on_val_batch_start()
        batch = transfer_batch_to_device(batch)
        val_step(batch)
        on_val_batch_end()
    end
    on_val_epoch_end()
end

Hooks API

Tsunami.on_before_backprop — Function

on_before_backprop([callback,] model, trainer, loss)

Called after the model's forward, where also the pullback is computed, but before the call to the pullback (the backward pass).

Tsunami.on_before_update — Function

on_before_update([callback,] model, trainer, grad)

Called before the call to Optimisers.update! that applies the gradient grad to update the model's parameters.

Tsunami.on_train_batch_start — Function

on_train_batch_start([callback,] model, trainer, batch, batch_idx)

Called at the beginning of each training batch.

Tsunami.on_train_batch_end — Function

on_train_batch_end([callback,] model, trainer)

Called at the end of each training batch.

Tsunami.on_train_epoch_start — Function

on_train_epoch_start([callback,] model, trainer)

Called at the beginning of each training epoch.

Tsunami.on_train_epoch_end — Function

on_train_epoch_end([callback,] model, trainer)

Called at the end of each training epoch.

Tsunami.on_test_batch_start — Function

on_test_batch_start([callback,] model, trainer, batch, batch_idx)

Called at the beginning of each test batch.

Tsunami.on_test_batch_end — Function

on_test_batch_end([callback,] model, trainer)

Called at the end of each test batch.

Tsunami.on_test_epoch_start — Function

on_test_epoch_start([callback,] model, trainer)

Called at the beginning of each test epoch.

Tsunami.on_test_epoch_end — Function

on_test_epoch_end([callback,] model, trainer)

Called at the end of each test epoch.

Tsunami.on_val_batch_start — Function

on_val_batch_start([callback,] model, trainer, batch, batch_idx)

Called at the beginning of each validation batch.

Tsunami.on_val_batch_end — Function

on_val_batch_end([callback,] model, trainer)

Called at the end of each validation batch.

Tsunami.on_val_epoch_start — Function

on_val_epoch_start([callback,] model, trainer)

Called at the beginning of each validation epoch.

Tsunami.on_val_epoch_end — Function

on_val_epoch_end([callback,] model, trainer)

Called at the end of each validation epoch.