Logging

Tsunami.log — Function

" log(trainer::Trainer, name::AbstractString, value; onstep = nothing, onepoch = nothing, prog_bar = false)

Log a value to the logger. Can be called from any function in the training loop or from a callback. Logs to the loggers specified in trainer.loggers.

See the Logging docs for more details.

Arguments

trainer::Trainer: The trainer object.

name::AbstractString: The name of the value.

value: The value to log.

on_step::Bool: Whether to log the value on each step. Defaults to true if stage is :train or :validate, false otherwise.

on_epoch::Bool: Whether to log the value on each epoch. Defaults to true if stage is :training_epoch_end or :validation_epoch_end, false otherwise.

prog_bar::Bool: Whether to log the value to the progress bar. Defaults to false.

Examples

function validation_step(model::Model, trainer, batch, batch_idx)
    # log the validation loss
    ...
    Tsunami.log(trainer, "val/loss", val_loss)
end