Hooks

Hooks are a way to extend the functionality of Tsunami. They are a way to inject custom code into the FluxModule or into a Callback at various points in the training, testing, and validation loops.

At a high level, and omitting function imputs and outputs, a simplified version of the Tsunami.fit! method looks like this:

function fit!()
    configure_optimizers()
    
    for epoch in epochs
        train_loop()
    end
end

function train_loop()
    set_learning_rate(lr_scheduler, epoch)

    for batch in train_dataloader
        batch = transfer_batch_to_device(batch)
        train_step(batch)
        compute_gradient()
        optimizer_step()
        if should_check_val
            val_loop()
        end
    end
    on_train_epoch_end()
end

function val_loop()
    for batch in val_dataloader
        batch = transfer_batch_to_device(batch)
        val_step(batch)
    end
    on_val_epoch_end()
end

Hooks API

Tsunami.on_test_epoch_end — Function

on_test_epoch_end(model, trainer)
on_test_epoch_end(callback, model, trainer)

Called in Tsunami.fit! at the end of each test epoch.

Tsunami.on_train_epoch_end — Function

on_train_epoch_end(model, trainer)
on_train_epoch_end(callback, model, trainer)

Called in Tsunami.fit! at the end of each training epoch.

Tsunami.on_val_epoch_end — Function

on_val_epoch_end(model, trainer)
on_val_epoch_end(callback, model, trainer)

Called in Tsunami.fit! at the end of each validation epoch.